+++ Sources/Sentry/SentryNetworkSwizzling.m
@@ -1,4 +1,5 @@
 #import "SentryNetworkSwizzling.h"
+#import "SentryHttpInterceptor+Private.h"
 #import "SentryNetworkTracker.h"
 #import "SentrySwizzle.h"
 
@@ -8,6 +9,12 @@
 {
     [SentryNetworkTracker.sharedInstance enable];
     [self swizzleURLSessionTaskResume];
+
+    // NSURLProtocol is only used when NSURLSession.shared is used.
+    // If a custom NSURLSession with custom configurations is used SentryHTTPProtocol is not called.
+    // To solve this we swizzle the NSURLSession session configuration and add SentryHTTPProtocol in
+    // the classProtocol list.
+    [self swizzleURLSessionWithConfiguration];
 }
 
 + (void)stop
@@ -30,5 +37,19 @@
         }),
         SentrySwizzleModeOncePerClassAndSuperclasses, (void *)selector);
 }
+
++ (void)swizzleURLSessionWithConfiguration
+{
+    SentrySwizzleClassMethod(NSURLSession.class,
+        NSSelectorFromString(@"sessionWithConfiguration:delegate:delegateQueue:"),
+        SentrySWReturnType(NSURLSession *),
+        SentrySWArguments(NSURLSessionConfiguration * configuration,
+            id<NSURLSessionDelegate> delegate, NSOperationQueue * queue),
+        SentrySWReplacement({
+            [SentryHttpInterceptor configureSessionConfiguration:configuration];
+            return SentrySWCallOriginal(configuration, delegate, queue);
+        }));
+}
+
 #pragma clang diagnostic pop
 @end