+++ CHANGELOG.md
@@ -3,11 +3,12 @@
 ## Unreleased
 
 - fix: Don't create transactions for HTTP Requests. (#1237)
+- feat: Add Sentry Trace HTTP Header (#1213)
 - feat: Sub-millis precision for spans and events (#1234)
 - fix: Race on session task (#1233)
 
 ## 7.2.0-beta.6
-
+ 
 - fix: ViewController swizzling before iOS 13 (#1231)
 - fix: AppStartMeasurement didFinishLaunching is nil (#1228)